requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.10.0
boto3==1.35.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
openai==1.54.0
pinecone-client==5.0.1
python-dotenv==1.0.1
tiktoken==0.9.0
mangum==0.19.0
bcrypt==4.1.2
starlette==0.41.0

